Use alternate image as fallback for creation date
This happens when the saved image is valid, but not present among the
generated images in the prompt URL.

This behavior is also present when accessing the saved image manually in
the collection via the browser. ie. The thumbnail shows the correct
image, but is not among the images in the prompt URL.

From initial testing, the creation date and prompt of the alternate
image, which is the first of the generated images in the prompt URL,
seems to line up with the saved one.