Use dateModified as fallback for creation date For reasons unknown, `images` under `set_creation_date()` may sometimes return `None` even when the original images are still valid. From personal observation, the likelihood of this occurring increases the older the saved generation gets.
This commit would use `dateModified` from a collection item as fallback
which, for all intents and purposes, would usually return the same date
as the creation date. It should also be just as if not more persistent
than the creation date because this date presumably applies to a
collection item as a whole, and I've yet to see an item completely
invalidated (as opposed to partially like missing thumbnail, original
image etc.).