Reverse the image list before downloading Currently, the indices are reversed compared with the API method.
Ideally, as with the API method, the first image should be the one at
the bottom right of the collection, which usually corresponds to the
oldest item.

But because of how the clipboard is normally copied with the "Select all
results in this collection" button, the first image ends up being the
topmost entry in the clipboard, corresponding to the top left item in
the collection, which is incorrect.

This commit would fix the issue and preserve naming parity with the API
method.